requirements.txt
dj-database-url==0.5.0
Django==2.2.16
django-ckeditor==6.0.0
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==8.0.1
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
psycopg2==2.7.2
